+++ packages/contracts/src/systems/_RegistryCreateModSystem.sol
@@ -7,16 +7,22 @@ import { getAddressById } from "solecs/utils.sol";
 
 import { LibRegistryItem } from "libraries/LibRegistryItem.sol";
 
-uint constant ID = uint(keccak256("system._Registry.Mod.Create"));
+uint256 constant ID = uint256(keccak256("system._Registry.Mod.Create"));
 
 // _RegistryCreateModSystem creates an item registry entry for a Mod item
 contract _RegistryCreateModSystem is System {
   constructor(IWorld _world, address _components) System(_world, _components) {}
 
   function execute(bytes memory arguments) public onlyOwner returns (bytes memory) {
-    (uint modIndex, string memory name, uint health, uint power, uint harmony, uint violence) = abi
-      .decode(arguments, (uint, string, uint, uint, uint, uint));
-    uint registryID = LibRegistryItem.getByModIndex(components, modIndex);
+    (
+      uint256 modIndex,
+      string memory name,
+      uint256 health,
+      uint256 power,
+      uint256 harmony,
+      uint256 violence
+    ) = abi.decode(arguments, (uint256, string, uint256, uint256, uint256, uint256));
+    uint256 registryID = LibRegistryItem.getByModIndex(components, modIndex);
 
     require(registryID == 0, "Item Registry: Mod index already exists");
 
@@ -25,12 +31,12 @@ contract _RegistryCreateModSystem is System {
   }
 
   function executeTyped(
-    uint modIndex,
+    uint256 modIndex,
     string memory name,
-    uint health,
-    uint power,
-    uint harmony,
-    uint violence
+    uint256 health,
+    uint256 power,
+    uint256 harmony,
+    uint256 violence
   ) public onlyOwner returns (bytes memory) {
     return execute(abi.encode(modIndex, name, health, power, harmony, violence));
   }
